fft/requirements-python.txt
# Python dependencies for PPM to JPG conversion script
Pillow>=9.0.0
python-dotenv>=0.19.0
